Opt-in to deploy:cleanup_assets with :keep_assets
In Rails 3, the `rake assets:clean` tasks deletes *all* assets, which is
obviously not the desired effect. To be safe, disable this task by default.

In Rails 4, `rake assets:clean` keeps the most recent 2 versions of assets
by default, and can be further tweaked by specifying a number, like
`rake assets:clean[5]` to keep 5 versions. To opt-in to this behavior, use
e.g. `set :keep_assets, 5`. This will enable the cleanup task and pass the `5`
value to rake.