Factor manifest backup/restore logic to Ruby class
Also improve and document the logic so that it is no longer dependent on `ls`
to find manifest files, which is fragile and is known to fail in cases where
more than one manifest file matches.

Structure the code in such a way that its behavior can be tested without
requiring Rake and a full-blown Capistrano and remote SSH server environment.

Add configuration options for `:asset_manifest_backup_directory` and
`:asset_manifest_patterns`.