Test manifest using local shell as fake SSH server
A `test/fixtures` directory contains various simulated Capistrano release
directories with Sprockets manifests and backups. During a test, the fixture is
copied to a temporary area where it can be safely operated on and modified
without affecting the original template

A fake Capistrano DSL is used so that we have access to `release_path`, `set`,
and `fetch` that a real Capistrano environment would provide. Instead of SSH,
commands are executed using a local shell using Ruby's `system`.

In this way we can test the backup and restore behavior and be reasonably sure
that they will also work in a real-world Capistrano deploy.